requirements.txt
PySide6
selenium
beautifulsoup4
torch
torchvision
opencv-python-headless
numpy
pillow
